WIP. Move code around, refactor, create codecs.
